ExecutionSettings

An R6 class to define an ExecutionSettings object

Active bindings

cdmDatabaseSchema

the schema containing the OMOP CDM

workDatabaseSchema

the schema containing the cohort table

tempEmulationSchema

the schema needed for temp tables

cohortTable

the table containing the cohorts

cdmSourceName

the name of the source data of the cdm

Methods

Public methods

ExecutionSettings$new()

ExecutionSettings$getDbms()

ExecutionSettings$connect()

ExecutionSettings$disconnect()

ExecutionSettings$getConnection()

ExecutionSettings$clone()

Method new()

Usage

ExecutionSettings$new(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema = NULL,
  workDatabaseSchema = NULL,
  tempEmulationSchema = NULL,
  cohortTable = NULL,
  cdmSourceName = NULL
)

Arguments

connectionDetails

a connectionDetails object

connection

a connection to a dbms

cdmDatabaseSchema

The schema of the OMOP CDM database

workDatabaseSchema

The schema to which results will be written

tempEmulationSchema

Some database platforms like Oracle and Snowflake do not truly support temp tables. To emulate temp tables, provide a schema with write privileges where temp tables can be created.

cohortTable

The name of the table where the cohort(s) are stored

cdmSourceName

A human-readable name for the OMOP CDM source

Method getDbms()

extract the dbms dialect

Usage

ExecutionSettings$getDbms()

Method connect()

connect to dbms

Usage

ExecutionSettings$connect()

Method disconnect()

disconnect from dbms

Usage

ExecutionSettings$disconnect()

Method getConnection()

retrieve the connection object

Usage

ExecutionSettings$getConnection()

Method clone()

The objects of this class are cloneable with this method.

Usage

ExecutionSettings$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.